rest api done

+++ controllers/note.js
@@ -84,6 +84,9 @@ exports.deleteNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
     .then((note) => {
+      if (note.author.toString() !== req.userId) {
+        return res.status(401).json({ message: "Auth failed!" });
+      }
       if (note.cover_image) {
         unlink(note.cover_image);
       }
@@ -105,6 +108,9 @@ exports.getOldNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
     .then((note) => {
+      if (note.author.toString() !== req.userId) {
+        return res.status(401).json({ message: "Auth failed!" });
+      }
       return res.status(200).json(note);
     })
     .catch((err) => {
@@ -121,6 +127,9 @@ exports.updateNote = (req, res, next) => {
 
   Note.findById(note_id)
     .then((note) => {
+      if (note.author.toString() !== req.userId) {
+        return res.status(401).json({ message: "Auth failed!" });
+      }
       note.title = title;
       note.content = content;
       if (cover_image) {
